Move supportText() to utils in order to allow use outside of errors
